dom/tests/mochitest/pointerlock/file_cursorPosEvents.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Bug 633602 - file_cursorPosEvents.html</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript" src="mouselock_util.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style type="text/css">
    #child {
      width: 100px;
      height: 100px;
      background-color:Green;
    }
  </style>
</head>
<body onload="start();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
  Mozilla Bug 633602</a>

  <div id="parent">
    <div id="child"></div>
  </div>

  <script type="application/javascript">
      /*
       * Test for Bug 633602
       * Test will check to make sure that the following mouse events are no
       * longer executed in pointer lock.
       *  - mouseover, mouseout, mouseenter, mouseleave
       */
      SimpleTest.waitForExplicitFinish();

      var parent, child;
      var pointer = navigator.mozPointer;

      function PointerEventStats() {
        this.mouseEnter = false;
        this.mouseLeave = false;
        this.mouseOver = false;
        this.mouseOut = false;
      }

      var parentStats = new PointerEventStats(),
        childStats = new PointerEventStats();

      document.addEventListener('mozfullscreenchange', function() {
        if (document.mozFullScreen && document.mozFullScreenElement === parent) {
          pointer.lock(parent,
            function() {
              parent.addEventListener('mousemove', runTests, false);
              synthesizeMouseAtCenter(child, { type: 'mousemove' }, window);
            },
            function () {
              ok(false, "Failed to lock, shouldn't happen.");
            }
          );
        } else {
          SimpleTest.finish();
        }
      }, false);

      function runTests() {
        parent.removeEventListener('mousemove', runTests, false);

        ok(pointer.isLocked, "expected mouse to be locked, but wasn't.");

        is(childStats.mouseEnter, false,
           "child's mouseenter should not be firing in Full Screen and Pointer Lock.");
        is(childStats.mouseOver, false,
           "child's mouseover should not be firing in Full Screen and Pointer Lock.");
        is(childStats.mouseLeave, false,
           "child's mouseleave should not be firing in Full Screen and Pointer Lock.");
        is(childStats.mouseOut, false,
           "child's mouseout should not be firing in Full Screen and Pointer Lock.");

        is(parentStats.mouseEnter, false,
           "parent's mouseenter should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.mouseOver, false,
           "parent's mouseover should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.mouseLeave, false,
           "parent's mouseleave should not be firing in Full Screen and Pointer Lock.");
        is(parentStats.mouseOut, false,
           "parent's mouseout should not be firing in Full Screen and Pointer Lock.");

        pointer.unlock();
      }

      function setupEventHandlers(element, stats) {
        element.addEventListener('mouseout', function() {
          stats.mouseOut = true;
        } , false);

        element.addEventListener('mouseleave', function() {
          stats.mouseLeave = true;
        } , false);

        element.addEventListener('mouseover', function() {
          stats.mouseOver = true;
        } , false);

        element.addEventListener('mouseenter', function() {
          stats.mouseEnter = true;
        } , false);
      }

      function start() {
        parent = document.getElementById('parent'),
        child = document.getElementById('child');

        parent.addEventListener('mozpointerlocklost', function() {
          document.mozCancelFullScreen();
        }, false);

        setupEventHandlers(parent, parentStats);
        setupEventHandlers(child, childStats);

        SimpleTest.waitForFocus(function() {
          parent.mozRequestFullScreen();
        });
      }
  </script>
</body>
</html>
